+++ src/gnu/usr.bin/lynx/src/LYPrettySrc.c
@@ -174,8 +174,8 @@ static void append_open_tag(char *tagname,
 #define isLeadP(p) ((isalpha(UCH(*p)) || *p == '_'))
 #define isNextP(p) ((isalnum(UCH(*p)) || *p == '_'))
 
-#define FMT_AT " at column %ld:\n\t%s\n"
-#define TXT_AT (1 + p - ts), ts
+#define FMT_AT " at column %d:\n\t%s\n"
+#define TXT_AT (int) (1 + p - ts), ts
 
 /* returns FALSE if incorrect */
 int html_src_parse_tagspec(char *ts,
